Fixes possible ioverflow by expanding the range returned.
Since some functions need to return a negative value, and Windows
uses 'ints' (which may or may not be 64 bits) for return values
of its functions, I made it all i64 all around.

The parameters are still u32, though, so the callers will not be
able to send more than a u32 in a single call.

+++ src/netcode_tcp.h
@@ -20,36 +20,42 @@ extern "C" {
     * it is allocated and filled with the IP address of the remote peer. If
     * port is not NULL then is it filled with the port of the remote peer.
     *
-    * On error -1 is returned. On timeout zero is returned. On success a non-zero
-    * file descriptor is returned.
+    * The returned value must be tested for failure using the macro
+    * NETCODE_SOCK_VALID(fd) which evaluates to true if 'fd' is a valid socket
+    * and evalutes to false if it is not. On timeout zero (an invalid socket)
+    * is returned.
     */
-   socket_t netcode_tcp_accept (socket_t fd, uint32_t timeout_secs, char **addr, uint16_t *port);
+   socket_t netcode_tcp_accept (socket_t fd, uint32_t timeout_secs,
+                                char **addr, uint16_t *port);
 
    /* Make a connection to the specified server on the specified port.
-    * On success the fd of the connected descriptor is returned. On error -1
-    * is returned.
+    * On success the socket of the connected descriptor is returned. Use
+    * the macro NETCODE_SOCK_VALID(fd) which evaluates to true is 'fd' is a
+    * valid socket.
     */
    socket_t netcode_tcp_connect (const char *server, uint16_t port);
 
-   /* Write the given buffer to the given fd. On success the number
+   /* Write the given buffer to the given socket. On success the number
     * of bytes written is returned, which may be less than the specified
     * number of bytes.
     *
-    * On error (uint32_t)-1 is returned.
+    * On error -1 is returned.
     */
-   uint32_t netcode_tcp_write (socket_t fd, const void *buf, uint32_t len);
+   int64_t netcode_tcp_write (socket_t fd, const void *buf, uint32_t len);
 
-   /* Read not more than the specified number of bytes from the given fd
+   /* Read not more than the specified number of bytes from the given socket
     * into the specified buffer. This function returns when the specified
     * buffer is full or when the timeout expires, whichever happens first.
     *
     * On success the number of bytes that were actually read is returned,
     * which will be less than or equal to the length specified. On error
-    * (uint32_t) -1 is returned.
+    * -1 is returned.
     *
     * No more than timeout seconds is spent filling the buffer.
     */
-   uint32_t netcode_tcp_read (socket_t fd, void *buf, uint32_t len, uint32_t timeout);
+   int64_t netcode_tcp_read (socket_t fd, void *buf, uint32_t len,
+                             uint32_t timeout);
+
 
 #ifdef __cplusplus
 };
